Retrieved API data,Stored in BufferedInputStream and Sent to XMLPullParser
I fixed the issue of not being able to retrieve data from API call by using a bufferedReader instead of an inputStream. I than read all lines out of the bufferedReader into one String. Than converted this string to a UTF-8.Than I was able to get an inputStream out of this and send it to the parser with all correct information.